requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1
python-multipart==0.0.6

# MQTT
paho-mqtt==1.6.1

# Face Recognition
face-recognition==1.3.0
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0

# Logging and monitoring
python-json-logger==2.0.7
